Added default DateTime value resolution
